Adding a simple test
