+++ tests/CMakeLists.txt
@@ -5,9 +5,9 @@ file(GLOB TEST_SRC "*.c")
 file(GLOB TEST_HDRS "*.h")
 file(GLOB TESTS ${TEST_HDRS} ${TEST_SRC})
 
-# Disabling tests for now until credentials are sorted out for CI.
 add_test_case(test_s3_client_create_destroy)
 add_test_case(test_s3_request_create_destroy)
+add_test_case(test_s3_pl_allocator)
 
 add_test_case(test_s3_meta_request_retry_queue_operations)
 add_test_case(test_s3_meta_request_retry_queue_clean_up)
