+++ tests/s3_misc_tests.c
@@ -1,0 +1,124 @@
+/**
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0.
+ */
+
+#include "aws/s3/private/s3_client_impl.h"
+#include "aws/s3/private/s3_meta_request_impl.h"
+#include "aws/s3/private/s3_util.h"
+#include "aws/s3/private/s3_pl_allocator.h"
+#include "s3_tester.h"
+#include <aws/common/byte_buf.h>
+#include <aws/common/clock.h>
+#include <aws/common/common.h>
+#include <aws/common/environment.h>
+#include <aws/common/ref_count.h>
+#include <aws/http/request_response.h>
+#include <aws/io/stream.h>
+#include <aws/io/tls_channel_handler.h>
+#include <aws/testing/aws_test_harness.h>
+#include <inttypes.h>
+
+AWS_TEST_CASE(test_s3_client_create_destroy, s_test_s3_client_create_destroy)
+static int s_test_s3_client_create_destroy(struct aws_allocator *allocator, void *ctx) {
+    (void)ctx;
+
+    struct aws_s3_tester tester;
+    AWS_ZERO_STRUCT(tester);
+    ASSERT_SUCCESS(aws_s3_tester_init(allocator, &tester));
+
+    struct aws_s3_client_config client_config;
+    AWS_ZERO_STRUCT(client_config);
+
+    ASSERT_SUCCESS(aws_s3_tester_bind_client(
+        &tester, &client_config, AWS_S3_TESTER_BIND_CLIENT_REGION | AWS_S3_TESTER_BIND_CLIENT_SIGNING));
+
+    struct aws_s3_client *client = aws_s3_client_new(allocator, &client_config);
+
+    ASSERT_TRUE(client != NULL);
+
+    aws_s3_client_release(client);
+    client = NULL;
+
+    aws_s3_tester_clean_up(&tester);
+
+    return 0;
+}
+
+AWS_TEST_CASE(test_s3_request_create_destroy, s_test_s3_request_create_destroy)
+static int s_test_s3_request_create_destroy(struct aws_allocator *allocator, void *ctx) {
+    (void)ctx;
+
+    const int request_tag = 1234;
+    const uint32_t part_number = 5678;
+
+    struct aws_s3_tester tester;
+    AWS_ZERO_STRUCT(tester);
+    ASSERT_SUCCESS(aws_s3_tester_init(allocator, &tester));
+
+    struct aws_s3_meta_request *meta_request = aws_s3_tester_meta_request_new(&tester, NULL, NULL);
+    ASSERT_TRUE(meta_request != NULL);
+
+    struct aws_http_message *request_message = aws_s3_tester_dummy_http_request_new(&tester);
+    ASSERT_TRUE(request_message != NULL);
+
+    struct aws_s3_request *request =
+        aws_s3_request_new(meta_request, request_tag, part_number, AWS_S3_REQUEST_DESC_RECORD_RESPONSE_HEADERS);
+
+    ASSERT_TRUE(request != NULL);
+
+    ASSERT_TRUE(request->meta_request == meta_request);
+    ASSERT_TRUE(request->part_number == part_number);
+    ASSERT_TRUE(request->request_tag == request_tag);
+    ASSERT_TRUE(request->record_response_headers == true);
+
+    aws_s3_request_setup_send_data(request, request_message);
+
+    ASSERT_TRUE(request->send_data.message != NULL);
+    ASSERT_TRUE(request->send_data.response_headers == NULL);
+
+    request->send_data.response_headers = aws_http_headers_new(allocator);
+    ASSERT_TRUE(request->send_data.response_headers != NULL);
+
+    aws_s3_request_clean_up_send_data(request);
+
+    ASSERT_TRUE(request->send_data.message == NULL);
+    ASSERT_TRUE(request->send_data.response_headers == NULL);
+    ASSERT_TRUE(request->send_data.response_status == 0);
+
+    aws_s3_request_release(request);
+    aws_http_message_release(request_message);
+    aws_s3_meta_request_release(meta_request);
+
+    aws_s3_tester_clean_up(&tester);
+
+    return 0;
+}
+
+AWS_TEST_CASE(test_s3_pl_allocator, s_test_s3_pl_allocator)
+static int s_test_s3_pl_allocator(struct aws_allocator *allocator, void *ctx) {
+    (void)ctx;
+
+    const size_t buffer_test_size = 1024;
+
+    struct aws_s3_tester tester;
+    AWS_ZERO_STRUCT(tester);
+    ASSERT_SUCCESS(aws_s3_tester_init(allocator, &tester));
+
+    struct aws_allocator *s3_pl_allocator = aws_s3_pl_allocator_new(allocator);
+
+    {
+        void *buffer = aws_mem_acquire(s3_pl_allocator, buffer_test_size);
+        aws_mem_release(s3_pl_allocator, buffer);
+    }
+    {
+        aws_s3_pl_allocator_acquire(s3_pl_allocator);
+        aws_s3_pl_allocator_release(s3_pl_allocator);
+    }
+
+    aws_s3_pl_allocator_release(s3_pl_allocator);
+
+    aws_s3_tester_clean_up(&tester);
+
+    return 0;
+}
